requirements.txt
customtkinter>=5.2.0
python-snap7>=1.3
mariadb>=1.1.0
pandas>=2.0.0